bootstrapper: allow building without cgo dependencies for linting
