+++ bootstrapper/internal/diskencryption/BUILD.bazel
@@ -3,13 +3,24 @@ load("//bazel/go:go_test.bzl", "go_ld_test", "go_test")
 
 go_library(
     name = "diskencryption",
-    srcs = ["diskencryption.go"],
+    srcs = [
+        "diskencryption.go",
+        "diskencryption_cgo.go",
+        "diskencryption_cross.go",
+    ],
     importpath = "github.com/edgelesssys/constellation/v2/bootstrapper/internal/diskencryption",
     visibility = ["//bootstrapper:__subpackages__"],
-    deps = [
-        "@com_github_martinjungblut_go_cryptsetup//:go-cryptsetup",
-        "@com_github_spf13_afero//:afero",
-    ],
+    deps = select({
+        "@io_bazel_rules_go//go/platform:android": [
+            "@com_github_martinjungblut_go_cryptsetup//:go-cryptsetup",
+            "@com_github_spf13_afero//:afero",
+        ],
+        "@io_bazel_rules_go//go/platform:linux": [
+            "@com_github_martinjungblut_go_cryptsetup//:go-cryptsetup",
+            "@com_github_spf13_afero//:afero",
+        ],
+        "//conditions:default": [],
+    }),
 )
 
 go_test(
@@ -18,12 +29,21 @@ go_test(
     embed = [":diskencryption"],
     # keep
     tags = ["manual"],
-    deps = [
-        "@com_github_spf13_afero//:afero",
-        "@com_github_stretchr_testify//assert",
-        "@com_github_stretchr_testify//require",
-        "@org_uber_go_goleak//:goleak",
-    ],
+    deps = select({
+        "@io_bazel_rules_go//go/platform:android": [
+            "@com_github_spf13_afero//:afero",
+            "@com_github_stretchr_testify//assert",
+            "@com_github_stretchr_testify//require",
+            "@org_uber_go_goleak//:goleak",
+        ],
+        "@io_bazel_rules_go//go/platform:linux": [
+            "@com_github_spf13_afero//:afero",
+            "@com_github_stretchr_testify//assert",
+            "@com_github_stretchr_testify//require",
+            "@org_uber_go_goleak//:goleak",
+        ],
+        "//conditions:default": [],
+    }),
 )
 
 go_ld_test(
